+++ packages/base/base/src/index.ts
@@ -1,2 +1,5 @@
-export * from './emitter'
+export * from './emitter/base'
+export * from './emitter/emitter'
+export * from './emitter/protected-scoped'
+export * from './emitter/protected'
 export * from './uuid'
